test(player): split player specs by lifecycle, audio, and playlist behavior
